Fixed some issues

+++ libs/cocos2d/CCSprite.m
@@ -502,10 +502,10 @@ struct transformValues_ {
 #define kQuadSize sizeof(quad_.bl)
 	glBindTexture(GL_TEXTURE_2D, [texture_ name]);
 	
-	int offset = (int)&quad_;
+	long offset = (long)&quad_;
 	
 	// vertex
-	int diff = offsetof( ccV3F_C4B_T2F, vertices);
+	NSInteger diff = offsetof( ccV3F_C4B_T2F, vertices);
 	glVertexPointer(3, GL_FLOAT, kQuadSize, (void*) (offset + diff) );
 	
 	// color
